+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.3] - 2025-09-15
+- Converted to use `settings.json` for EverQuest and Obsidian vault paths
+- Added first-run GUI prompt for selecting directories
+
 ## [0.1.2] - 2025-09-08
 - Added Global Inventory summary and Velious gems summary